Refactor code due to code review
